requirements.txt
python-telegram-bot
python-dotenv~=1.0.1
requests~=2.32.3
schedule~=1.2.2